workspace/projects/reo_mcp/requirements.txt
# Reo MCP server — pinned production dependencies.
# Pin exact versions per §1.1 of reo_pinata_agent_context.md.

fastmcp==3.2.4
httpx==0.28.1
python-dotenv==1.2.2

# Dev / test (kept with prod for a single-install flow on Pinata).
# If the container's install layer ever cares about size, split into
# requirements-dev.txt.
pytest==9.0.3
pytest-asyncio==1.3.0
respx==0.23.1
ruff==0.15.10
